+++ test/account/AccountECDSA.test.js
@@ -48,7 +48,6 @@ describe('AccountECDSA', function () {
   describe('ERC7739', function () {
     beforeEach(async function () {
       this.mock = await this.mock.deploy();
-      this.signTypedData = this.signer.signTypedData.bind(this.signer);
     });
 
     shouldBehaveLikeERC7739();